+++ src/speech/google/google-module-config.interface.ts
@@ -19,4 +19,8 @@
 export interface GoogleModuleConfigInterface {
     /** legt den AppKey fuer die Verbindung zum Server fest */
     googleAppKey: string;
+    /** legt die URL fuer den Dialogflow Tokenserver fest */
+    dialogflowTokenServerUrl: string;
+    /** legt die   ProjektID fuer Dialogflow fest */
+    dialogflowProjectId: string;
 }