+++ src/speech/google/google-service.spec.ts
@@ -63,9 +63,11 @@ describe('GoogleService', () => {
     describe('Funktion setCredentials', () => {
 
         it('sollte Credentials eintragen', () => {
-            expect( googleService.setCredentials( 'TestAppKey' )).toBe( 0 );
+            expect( googleService.setCredentials( 'TestAppKey', 'TestTokenServerUrl', 'TestProjectId' )).toBe( 0 );
             const credentials = GoogleModule.getConfig();
             expect( credentials.googleAppKey ).toBe( 'TestAppKey' );
+            expect( credentials.dialogflowTokenServerUrl ).toBe( 'TestTokenServerUrl' );
+            expect( credentials.dialogflowProjectId ).toBe( 'TestProjectId' );
         });
 
     });