+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Speech-React Release Notizen
 
 
+## 0.5.15.0008 Beta (19.10.2019)
+
+### Features
+
+* **Dialog:** Erweiterung der Dialog-Komponente um das Einlesen von JSON-Daten anstelle einer Def-Datei.
+* **Google:** Erweiterung der Google-Komponente um Dialogflow Version 2.
+
+
 ## 0.5.14.0007 Beta (31.08.2019)
 
 ### Features